+++ src/xpdev/sockwrap.h
@@ -228,8 +228,8 @@ extern "C" {
 DLLEXPORT socket_option_t* getSocketOptionList(void);
 DLLEXPORT int getSocketOptionByName(const char* name, int* level);
 
-DLLEXPORT int sendfilesocket(int sock, int file, off_t* offset, off_t count);
-DLLEXPORT int recvfilesocket(int sock, int file, off_t* offset, off_t count);
+DLLEXPORT off_t sendfilesocket(int sock, int file, off_t* offset, off_t count);
+DLLEXPORT off_t recvfilesocket(int sock, int file, off_t* offset, off_t count);
 DLLEXPORT BOOL socket_check(SOCKET sock, BOOL* rd_p, BOOL* wr_p, DWORD timeout);
 DLLEXPORT int retry_bind(SOCKET s, const struct sockaddr *addr, socklen_t addrlen
 				   ,uint retries, uint wait_secs, const char* prot
